feat: strongly-typed http.Handler builder
Creates a builder for strongly-typed http.Handler implementations and a
few minimal convenience functions. The current design is entirely
focused on the mapping from HTTP requests to strongly-typed functions,
and from their return values back to HTTP responses. The priority at
this point is to keep the API as composable as possible and avoid
introducing anything opinionated (besides the fact that we want strong
typing) and avoid any dependencies that would discourage
interoperability with web frameworks and libraries that have already
solved problems we haven't or don't intend to.