feat: switch Route from Handler to HandlerFunc
Updates the Route function to return an http.HandlerFunc instead of an
http.Handler. The http.HandlerFunc type implements http.Handler so
returning the allows the result to be consumed as either type directly
where a user would previously have had to call .ServeHTTP on the
returned http.Handler if they wanted an http.HandlerFunc. Internally ths
implementation uses a type implementing http.Handler with the intention
that we might have wanted to extend the capability of the marshaling
logic, but if not we should benchmark the current implementation VS
directly returning closures and make the choice that way.