feat!: add http.Request to Responder interface
The original implementation and interfaces represent the process of
performing business logic within an HTTP request handler as three steps:
parse the request, perform the logic, write the response. However, it's
actually more like a specific category of middleware where the outer
layer handles translation to and from HTTP and the inner (next) layer
performs the logic. The parser is the part of the translation layer that
happens before the inner (next) layer is invoked and the response writer
is the part of the outer layer that happens after the inner (next) layer
has returned. We want to hide the HTTP details from the inner layer but
it makes sense for the outer layer, including the response writer, to
have access to it all. For example, the response writer may be
interested in request headers about response encoding.

In an ideal representation we would provide a strongly typed way to
share the context from the parser with the response writer to strengthen
the concept of them being the before/after components of the same
middleware layer. Unfortunately, we haven't figured out a way to write
it as a single function because we can't force that function to call the
Target in the middle, nor to even include strongly-typed representations
within its own body. For now we'll take the stance that we lean on the
http.Request and custom context.Context values.