wip: example DSL for DI and app building
Shows an example DSL for registering services for DI and building a
middleware pipeline. This doesn't show endpoint building but it would
basically involve an Endpoint function -- sadly a free function because
it needs to be generic -- taking a route pattern, request parser, target
function, and a response writer. The endpoint object would have methods
for adding middlware like the app except they would apply only to the
endpoint.

Also considering an EndpointGroup that we could register middlware on
that would only belong to the group instead of the whole app, then add
endpoints to the group. Likely EndpointGroup would be an interface and
App would implement it so we can just implement Endpoint registration on
the EndpointGroup interface.

Similarly considering a MiddlewareGroup that would simply aggregate
middleware builders so that multiple middleware can be added in a single
line anywhere middleware can be added.