feat: creates skeleton API for inject package
Creates the intended public API for the inject package. Nearly all of
the intended functionality is unimplemented at this point. The
intentions of publishing the non-functional API are to have a concrete
basis for discussion about the API's intention and implementation, and
to support leaning on the compiler to test the ergonomics of the API.