feat!: scoped services + func registration
Adds support for scoped services where the same instance is returned
every time a service type is requested from the same provider. The
intention for singletons is that a "root" provider will own the
singletons and child scopes will reach back to their root for them, but
scopes created from other scopes will be independent and provide their
own scoped services.

Adds an implementation for RegisterFunc that uses two type parameters --
one for the service type and one for the implementation type -- and uses
a runtime check to ensure the implmentation type is assignable to the
service type. This is necessary because Go doesn't allow type parameters
to be used as constraints (e.g. [Service any, Impl Service]). We could
have used Service for the factory function too since the impl type would
be returnable, but then we wouldn't be able to determine the return type
without calling the function (and even then it wouldn't be 100% reliable
since a function could return different types) which would interfere
with dependency graph analysis when we get there.

Removes the TypeOf function in favour of the TypeFor function from the
relfect package.

Improves some local/receiver variable names.