feat: implement singletons
Singletons are initialized once and the same instance is returned for
every resolution. For a single provider this is identical to scoped
instances, but when a new scope is created from another provider the
singletons will be shared by all scopes descendend from the same root.

NOTE that Scopes are not implemented yet.