chore: POC of app using controllers
Creates a POC of an app using controllers. The POC puts a barebones
app-host framework in the package -- even so much that it doesn't care
about strong typing -- and implements the endpoint builders in the app
itself. This may or may not be desirable for the final product. An app
host that simply knows how to build HTTP pipelines is a nice idea, and
strongly typed endpoint builders could be a separate package (or even
module), but we still have the desire to produce top-level constructs
like swagger docs using the strong types. With that in mind, we would
need to consider a solution that allows enumeration of the whole API to
notice the types even when the framework defining the structure of the
API doesn't know whether endpoints will have them or not.

Also removes a load of things that were moved to garlic.